+++ meet/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "cookie-parser": "^1.4.6",
     "express": "^4.18.2",
-    "peerjs": "^1.3.2",
+    "peerjs": "^1.4.7",
     "socket.io": "^4.5.4",
     "socket.io-client": "^4.5.4",
     "uuid": "^9.0.0"
